requirements.txt
appdata==2.1.2
asgiref==3.4.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.0
boto3==1.21.13
botocore==1.24.13
certifi==2021.5.30
cffi==1.15.0
click==8.0.3
colored==1.4.3
cryptography==36.0.2
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.10.1
django-countries==7.3.2
django-rest-knox==4.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.1.0
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.1
gunicorn==20.1.0
hrm==0.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
Pillow==8.4.0
pipenv==2021.5.29
platformdirs==2.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pymongo==3.12.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pyzk==0.9
s3transfer==0.5.2
semantic-version==2.9.0
setuptools-rust==1.1.2
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.2
to-requirements.txt==1.3.0
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.7.2
virtualenv-clone==0.5.6
Werkzeug==2.0.1
